# specs/models/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MLHub Models API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/models-api/artifacts/{artifact_id}/model": {
      "post": {
        "tags": [
          "Artifacts"
        ],
        "summary": "Associate an owned model with an artifact",
        "operationId": "associate_model_with_artifact",
        "parameters": [
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociateModelBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Model associated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssociateModelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Model or artifact not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          },
          "409": {
            "description": "Artifact already associated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to associate model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/models-api/external-models/{external_model_id}/deployment-options": {
      "get": {
        "tags": [
          "ExternalModels"
        ],
        "summary": "List deployment options for an external model",
        "operationId": "list_external_model_deployment_options",
        "parameters": [
          {
            "name": "external_model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of deployment options to return. Values are capped at 100.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned by the previous page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_count",
            "in": "query",
            "description": "Include the number of deployment options for the ExternalModel.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment options listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListExternalModelDeploymentOptionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid external model identifier or cursor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "ExternalModel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to list deployment options",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/models-api/external-models/{external_model_id}/deployment-options/{deployment_option_id}": {
      "get": {
        "tags": [
          "ExternalModels"
        ],
        "summary": "Get a deployment option and its required parameters",
        "operationId": "get_external_model_deployment_option",
        "parameters": [
          {
            "name": "external_model_id",
            "in": "path",
            "description": "ExternalModel identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "deployment_option_id",
            "in": "path",
            "description": "DeploymentOption identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment option found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetExternalModelDeploymentOptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "ExternalModel or DeploymentOption not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to get deployment option",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/models-api/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List owned or shared models",
        "operationId": "list_models",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "Owned",
                "Shared"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include_count",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Models listed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListModelsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to list models"
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Add an external model to the user's collection",
        "operationId": "create_model",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateModelBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Model created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateModelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "ExternalModel not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponse"
                }
              }
            }
          },
          "409": {
            "description": "Model already in your collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unable to create model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssociateModelBody": {
        "type": "object",
        "required": [
          "model_id"
        ],
        "properties": {
          "model_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "AssociateModelResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {
            "$ref": "#/components/schemas/Model"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
      "BadRequestResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {},
          "status": {
            "type": "integer",
            "format": "int32",
            "default": 400,
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
      "ConflictResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {},
          "status": {
            "type": "integer",
            "format": "int32",
            "default": 409,
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
      "CreateModelBody": {
        "type": "object",
        "required": [
          "name",
          "external_model_id"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "external_model_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        },
        "additionalProperties": false
      },
      "CreateModelResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {
            "$ref": "#/components/schemas/Model"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "GetExternalModelDeploymentOptionResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {
            "$ref": "#/components/schemas/DeploymentOptionDetail"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "ListExternalModelDeploymentOptionsResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentOption"
            }
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "ListModelsResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            }
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "Model": {
        "type": "object",
        "required": [
          "id",
          "name",
          "tenant_id",
          "owner",
          "external_model_id",
          "external_model",
          "visibility",
          "updated_at",
          "created_at"
        ],
        "properties": {
          "artifact_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "external_model": {
            "$ref": "#/components/schemas/ExternalModel"
          },
          "external_model_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/Visibility"
          }
        }
      },
...
      "NotFoundResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {},
          "status": {
            "type": "integer",
            "format": "int32",
            "default": 404,
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "ServerErrorResponse": {
        "type": "object",
        "required": [
          "result",
          "status",
          "message",
          "metadata",
          "version"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "result": {},
          "status": {
            "type": "integer",
            "format": "int32",
            "default": 500,
            "minimum": 0
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "Visibility": {
        "type": "string",
        "enum": [
          "Public",
          "Private"
        ]
      }
    }
  }
}